pyproject.toml
[project]
name = "cbs-bot"
version = "0.1.0"
description = "Discord bot: CluesBySam scores to Google Sheets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
